pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vzhtelegram"
version = "0.1.0"
description = "Automated Telegram channel pipeline: monitoring, scoring, generation, publishing"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.9,<4",
    "feedparser>=6.0,<7",
    "aiosqlite>=0.20,<1",
    "aiogram>=3.4,<4",
    "openai>=1.30,<2",
    "structlog>=24.1,<25",
    "tenacity>=8.2,<9",
    "pyyaml>=6.0,<7",
    "beautifulsoup4>=4.12,<5",
    "lxml>=5.1,<6",
    "apscheduler>=3.10,<4",
    "datasketch>=1.6,<2",
    "python-dateutil>=2.9,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "aioresponses>=0.7",
    "ruff>=0.4",
]

[project.scripts]
vzh-pipeline = "scripts.run_pipeline:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
